+++ src/core/api/movie-api.ts
@@ -2,7 +2,7 @@ import {createApi} from '@reduxjs/toolkit/query/react';
 import {env} from '../../env';
 import {ListResponse, Movie} from './types';
 import {dynamicBaseQuery} from './base-query';
-import {sanitizeQuery} from '../utils/sanitizeQuery';
+import {sanitizeQuery} from '../utils/sanitize-query';
 
 type MovieListResponse = ListResponse<Movie>;
